perf(ci): split 5 more heavy test files + increase to 40 shards
