+++ tests/worker-bootstrap-multi-learner-regression-setup.test.js
@@ -696,336 +696,3 @@ test('multi-learner #6: viewer learner D excluded from everything', async () => 
   }
 });
 
-// #7 — Edge case: preferredLearnerId switches selected learner.
-test('multi-learner #7: preferredLearnerId switches selected learner', async () => {
-  const server = createServer();
-  try {
-    seed4LearnerFixture(server);
-
-    const response = await postBootstrap(server, { preferredLearnerId: 'learner-b' });
-    assert.equal(response.status, 200);
-    const payload = await readJsonBody(response);
-    assert.equal(payload.ok, true);
-
-    // B becomes selected.
-    assert.equal(payload.account?.selectedLearnerId, 'learner-b',
-      'B is selected after preferredLearnerId');
-    assert.equal(payload.learners?.selectedId, 'learner-b',
-      'learners.selectedId reflects B');
-
-    // B gets full first-paint data (sessions/events bounded to B).
-    // B was seeded with 2 sessions and 20 events — guard against vacuous-truth.
-    assert.equal(payload.practiceSessions.length, 2, 'B has 2 sessions');
-    assert.equal(
-      payload.practiceSessions.every((s) => s.learnerId === 'learner-b'),
-      true,
-      'sessions bounded to B after switch',
-    );
-    assert.equal(payload.eventLog.length, 20, 'B has 20 events');
-    assert.equal(
-      payload.eventLog.every((e) => e.learnerId === 'learner-b'),
-      true,
-      'events bounded to B after switch',
-    );
-
-    // A and C appear as compact siblings.
-    const siblingIds = (payload.account?.learnerList || []).map((e) => e.id).sort();
-    assert.deepEqual(siblingIds, ['learner-a', 'learner-c'],
-      'A and C appear as compact siblings when B is selected');
-
-    // Subject states still include ALL writable learners.
-    const subjectKeys = Object.keys(payload.subjectStates || {}).sort();
-    assert.equal(subjectKeys.filter((k) => k.startsWith('learner-a')).length, 3, 'A subject states present');
-    assert.equal(subjectKeys.filter((k) => k.startsWith('learner-b')).length, 3, 'B subject states present');
-    assert.equal(subjectKeys.filter((k) => k.startsWith('learner-c')).length, 3, 'C subject states present');
-  } finally {
-    server.close();
-  }
-});
-
-// #8 — Edge case: preferredLearnerId pointing at viewer falls back.
-test('multi-learner #8: preferredLearnerId pointing at viewer falls back to persisted writable learner', async () => {
-  const server = createServer();
-  try {
-    seed4LearnerFixture(server);
-
-    // Point at viewer D — should silently fall back.
-    const response = await postBootstrap(server, { preferredLearnerId: 'learner-d' });
-    assert.equal(response.status, 200);
-    const payload = await readJsonBody(response);
-    assert.equal(payload.ok, true);
-
-    // Falls back to persisted writable learner (learner-a was selected in seed).
-    assert.equal(payload.account?.selectedLearnerId, 'learner-a',
-      'falls back to learner-a when preferred is viewer');
-    assert.equal(payload.learners?.selectedId, 'learner-a');
-
-    // Verify the data envelope is populated, not just metadata.
-    assert.ok(Object.keys(payload.subjectStates || {}).length >= 3,
-      'viewer fallback: subject states populated');
-    assert.equal(payload.practiceSessions.length, 5,
-      'viewer fallback envelope includes A sessions');
-
-    // No error.
-    assert.equal(payload.error, undefined, 'no error returned');
-    assert.equal(payload.code, undefined, 'no error code returned');
-
-    // Exercise true alphabetical fallback: clear the persisted selection
-    // so neither preferredLearnerId nor selected_learner_id resolve.
-    runSql(server, 'UPDATE adult_accounts SET selected_learner_id = NULL WHERE id = ?', [ACCOUNT_ID]);
-    const fallbackRes = await postBootstrap(server, { preferredLearnerId: 'learner-d' });
-    assert.equal(fallbackRes.status, 200);
-    const fallbackPayload = await readJsonBody(fallbackRes);
-    assert.equal(fallbackPayload.ok, true);
-    assert.equal(fallbackPayload.account?.selectedLearnerId, 'learner-a',
-      'viewer + no persisted selection falls back to alphabetical first writable');
-  } finally {
-    server.close();
-  }
-});
-
-// #9 — Edge case: cold-start alphabetical selection.
-test('multi-learner #9: cold-start alphabetical selection (no preferredLearnerId, no persisted)', async () => {
-  const server = createServer();
-  try {
-    // Seed WITHOUT selecting any learner (no `selected: true`).
-    insertLearner(server, { id: 'learner-a', name: 'Alpha', sortIndex: 0, role: 'owner' });
-    insertLearner(server, { id: 'learner-b', name: 'Beta', sortIndex: 1, role: 'member' });
-    insertLearner(server, { id: 'learner-c', name: 'Gamma', sortIndex: 2, role: 'member' });
-    insertSubjectState(server, 'learner-a', 'spelling');
-    insertSubjectState(server, 'learner-b', 'grammar');
-
-    const response = await postBootstrap(server, {});
-    assert.equal(response.status, 200);
-    const payload = await readJsonBody(response);
-    assert.equal(payload.ok, true);
-
-    // Alphabetical first writable learner selected.
-    assert.equal(payload.account?.selectedLearnerId, 'learner-a',
-      'cold-start: alphabetical first writable learner selected');
-    assert.equal(payload.learners?.selectedId, 'learner-a');
-
-    // No practice sessions were seeded for any learner in this cold-start
-    // fixture — assert explicitly empty rather than relying on vacuous `[].every()`.
-    assert.equal(payload.practiceSessions.length, 0,
-      'cold-start: zero sessions seeded, array is empty-but-correct');
-
-    // Subject states must ship for writable learners — the PR #316 regression
-    // was invisible without inspecting this envelope in cold-start scenarios.
-    const subjectKeys = Object.keys(payload.subjectStates || {});
-    assert.ok(subjectKeys.length >= 2, 'cold-start: subject states ship for writable learners');
-    // Check that sibling (learner-b) subject state is present, not just selected (learner-a).
-    assert.ok(subjectKeys.some((k) => k.startsWith('learner-b')),
-      'cold-start: sibling learner-b subject state present');
-  } finally {
-    server.close();
-  }
-});
-
-// #10 — Integration: notModified invalidation on sibling subject-state write.
-test('multi-learner #10: notModified invalidation on sibling subject-state write', async () => {
-  const server = createServer();
-  try {
-    seed4LearnerFixture(server);
-
-    // Get baseline hash H1.
-    const baseline = await postBootstrap(server, {});
-    const basePayload = await readJsonBody(baseline);
-    assert.equal(basePayload.ok, true);
-    const H1 = basePayload.revision?.hash;
-    assert.match(H1, /^[0-9a-f]{32}$/, 'baseline hash present');
-
-    // Confirm notModified on same hash.
-    const unchanged = await postBootstrap(server, { lastKnownRevision: H1 });
-    const unchangedPayload = await readJsonBody(unchanged);
-    assert.equal(unchangedPayload.notModified, true,
-      'pre-mutation: matching hash returns notModified');
-
-    // Simulate sibling B subject-state write: bump state_revision + mutate data.
-    runSql(server, `
-      UPDATE learner_profiles
-      SET state_revision = state_revision + 1, updated_at = ?
-      WHERE id = ?
-    `, [NOW + 1, 'learner-b']);
-    runSql(server, `
-      UPDATE child_subject_state
-      SET data_json = ?, updated_at = ?
-      WHERE learner_id = ? AND subject_id = ?
-    `, [
-      JSON.stringify({
-        prefs: { mode: 'learn', marker: { fixture: 'learner-b-grammar-mutated', progress: 99 } },
-        mastery: {
-          concepts: {
-            sentence_functions: {
-              attempts: 3,
-              correct: 3,
-              wrong: 0,
-              strength: 0.9,
-              correctStreak: 3,
-              intervalDays: 8,
-              dueAt: NOW + 365 * 86_400_000,
-            },
-          },
-        },
-      }),
-      NOW + 1,
-      'learner-b',
-      'grammar',
-    ]);
-
-    // Re-post with old H1 — must get full bundle, NOT notModified.
-    const response = await postBootstrap(server, { lastKnownRevision: H1 });
-    assert.equal(response.status, 200);
-    const payload = await readJsonBody(response);
-    assert.notEqual(payload.notModified, true,
-      'sibling subject-state write must invalidate notModified');
-    assert.notEqual(payload.revision?.hash, H1,
-      'new hash differs from H1 after sibling write');
-
-    // Value-level: mutated data is reflected in the redacted read-model.
-    const bGrammar = payload.subjectStates?.['learner-b::grammar'];
-    assert.ok(bGrammar, 'B grammar state present in full bundle');
-    assert.deepEqual(bGrammar.data, {}, 'B grammar raw data stays stripped');
-    assert.equal(bGrammar?.ui?.prefs?.marker?.fixture, 'learner-b-grammar-mutated',
-      'mutated fixture marker ships through the read-model');
-    assert.equal(bGrammar?.ui?.prefs?.mode, 'learn',
-      'mutated prefs mode ships through the read-model');
-    assert.equal(bGrammar?.ui?.stats?.concepts?.secured, 1,
-      'mutated mastery ships through the read-model stats');
-  } finally {
-    server.close();
-  }
-});
-
-// #11 — Integration: notModified invalidation on sibling game-state write.
-test('multi-learner #11: notModified invalidation on sibling game-state write', async () => {
-  const server = createServer();
-  try {
-    seed4LearnerFixture(server);
-
-    // Get baseline hash.
-    const baseline = await postBootstrap(server, {});
-    const basePayload = await readJsonBody(baseline);
-    const H1 = basePayload.revision?.hash;
-    assert.match(H1, /^[0-9a-f]{32}$/, 'baseline hash present');
-
-    // Confirm notModified.
-    const unchanged = await postBootstrap(server, { lastKnownRevision: H1 });
-    assert.equal((await readJsonBody(unchanged)).notModified, true, 'pre-mutation notModified');
-
-    // Simulate sibling C game-state write: bump state_revision.
-    runSql(server, `
-      UPDATE learner_profiles
-      SET state_revision = state_revision + 1, updated_at = ?
-      WHERE id = ?
-    `, [NOW + 1, 'learner-c']);
-    // Mutate C's game state: flip inklet branch from b1 to b2. The branch
-    // field survives the public transform pipeline, providing a reliable
-    // mutation witness.
-    runSql(server, `
-      UPDATE child_game_state
-      SET state_json = ?, updated_at = ?
-      WHERE learner_id = ? AND system_id = ?
-    `, [
-      JSON.stringify({
-        inklet: { mastered: ['learner-c-ink-word'], caught: true, branch: 'b2' },
-        glimmerbug: { mastered: ['learner-c-glim-word'], caught: true, branch: 'b1' },
-      }),
-      NOW + 1,
-      'learner-c',
-      'monster-codex',
-    ]);
-
-    // Re-post with old hash — must get full bundle.
-    const response = await postBootstrap(server, { lastKnownRevision: H1 });
-    assert.equal(response.status, 200);
-    const payload = await readJsonBody(response);
-    assert.notEqual(payload.notModified, true,
-      'sibling game-state write must invalidate notModified');
-    assert.notEqual(payload.revision?.hash, H1,
-      'new hash differs after sibling game-state write');
-
-    // Value-level: C's mutated game state ships through with flipped branch.
-    const cGame = payload.gameState?.['learner-c::monster-codex'];
-    assert.ok(cGame, 'C game state present in full bundle');
-    assert.equal(cGame?.inklet?.branch, 'b2',
-      'mutated inklet branch ships through (was b1, now b2)');
-  } finally {
-    server.close();
-  }
-});
-
-// #12 — Edge case: single-learner regression guard.
-test('multi-learner #12: single-learner regression guard', async () => {
-  const server = createServer();
-  try {
-    // Separate 1-learner fixture.
-    insertLearner(server, { id: 'learner-solo', name: 'Solo', sortIndex: 0, role: 'owner', selected: true });
-    insertSubjectState(server, 'learner-solo', 'spelling');
-    insertSubjectState(server, 'learner-solo', 'grammar');
-    insertGameState(server, 'learner-solo');
-    insertPracticeSession(server, 'learner-solo', { id: 'solo-sess-0' });
-    insertEvent(server, 'learner-solo', { id: 'solo-evt-0' });
-
-    const response = await postBootstrap(server, {});
-    assert.equal(response.status, 200);
-    const payload = await readJsonBody(response);
-    assert.equal(payload.ok, true);
-
-    // Subject states present.
-    const subjectKeys = Object.keys(payload.subjectStates || {}).sort();
-    assert.deepEqual(subjectKeys, ['learner-solo::grammar', 'learner-solo::spelling'],
-      'single-learner: both subject states present');
-
-    // Game state present.
-    const gameKeys = Object.keys(payload.gameState || {});
-    assert.equal(gameKeys.length, 1, 'single-learner: 1 game state entry');
-    assert.ok(payload.gameState['learner-solo::monster-codex'], 'single-learner: monster-codex present');
-
-    // Sessions and events present.
-    assert.equal(payload.practiceSessions.length, 1, 'single-learner: 1 session');
-    assert.equal(payload.eventLog.length, 1, 'single-learner: 1 event');
-
-    // subjectStatesBounded marker.
-    assert.equal(payload.bootstrapCapacity?.subjectStatesBounded, false,
-      'single-learner: subjectStatesBounded is false');
-
-    // Selected learner.
-    assert.equal(payload.account?.selectedLearnerId, 'learner-solo');
-    assert.equal(payload.account?.learnerList?.length, 0, 'single-learner: no siblings');
-  } finally {
-    server.close();
-  }
-});
-
-// #13 — Edge case: bootstrapCapacity.subjectStatesBounded marker.
-test('multi-learner #13: bootstrapCapacity.subjectStatesBounded marker for multi and single', async () => {
-  const server = createServer();
-  try {
-    seed4LearnerFixture(server);
-
-    // Multi-learner: subjectStatesBounded is false.
-    const multiResponse = await postBootstrap(server, {});
-    const multiPayload = await readJsonBody(multiResponse);
-    assert.equal(multiPayload.ok, true);
-    assert.equal(multiPayload.bootstrapCapacity?.subjectStatesBounded, false,
-      'multi-learner: subjectStatesBounded is false (states shipped for all writable)');
-  } finally {
-    server.close();
-  }
-
-  // Separate server for single-learner.
-  const soloServer = createServer();
-  try {
-    insertLearner(soloServer, { id: 'learner-solo', name: 'Solo', sortIndex: 0, role: 'owner', selected: true });
-    insertSubjectState(soloServer, 'learner-solo', 'spelling');
-
-    const soloResponse = await postBootstrap(soloServer, {});
-    const soloPayload = await readJsonBody(soloResponse);
-    assert.equal(soloPayload.ok, true);
-    assert.equal(soloPayload.bootstrapCapacity?.subjectStatesBounded, false,
-      'single-learner: subjectStatesBounded is also false');
-  } finally {
-    soloServer.close();
-  }
-});
